Revert "feat(ui): role and deployment-mode adaptive UI (landing, sidebar, top…"
This reverts commit 9d54a599b0e5cc6bf1ce75603229bfbb73a0af92.
